+++ examples/resources/host_mac_login_item/resource.tf
@@ -4,9 +4,5 @@ resource "host_package_brew" "hammerspoon" {
 }
 
 resource "host_mac_login_item" "hammerspoon" {
-  path = "/Applications/Hammerspoon.app"
-
-  depends_on = [
-    host_package_brew.hammerspoon,
-  ]
+  path = host_package_brew.hammerspoon.app_path
 }