+++ examples/resources/host_mac_setting/resource.tf
@@ -1,8 +1,5 @@
 resource "host_mac_setting" "dock_autohide" {
-  domain = {
-    apple = "dock"
-  }
-
-  key   = "autohide"
-  value = true
+  domain = "com.apple.dock"
+  key    = "autohide"
+  value  = true
 }